fix: resolve Ruff lint CI failures with ruff config and code fixes

+++ meteostat/providers/metno/forecast.py
@@ -1,4 +1,3 @@
-
 import pandas as pd
 from requests.exceptions import HTTPError
 
@@ -97,12 +96,8 @@ def map_data(record):
             transform=percentage_to_okta,
         ),
         Parameter.RHUM: safe_get(record, details_instant + ["relative_humidity"]),
-        Parameter.PRCP: safe_get(
-            record, details_next_1_hour + ["precipitation_amount"]
-        ),
-        Parameter.WSPD: safe_get(
-            record, details_instant + ["wind_speed"], transform=lambda x: x * 3.6
-        ),
+        Parameter.PRCP: safe_get(record, details_next_1_hour + ["precipitation_amount"]),
+        Parameter.WSPD: safe_get(record, details_instant + ["wind_speed"], transform=lambda x: x * 3.6),
         Parameter.WPGT: safe_get(
             record,
             details_instant + ["wind_speed_of_gust"],
@@ -111,11 +106,9 @@ def map_data(record):
         Parameter.WDIR: safe_get(
             record,
             details_instant + ["wind_from_direction"],
-            transform=lambda x: int(round(x)),
+            transform=lambda x: round(x),
         ),
-        Parameter.PRES: safe_get(
-            record, details_instant + ["air_pressure_at_sea_level"]
-        ),
+        Parameter.PRES: safe_get(record, details_instant + ["air_pressure_at_sea_level"]),
         Parameter.COCO: safe_get(
             record,
             ["data", "next_1_hours", "summary", "symbol_code"],
@@ -174,9 +167,7 @@ def get_df(latitude: float, longitude: float, elevation: int) -> pd.DataFrame | 
 
     except HTTPError as error:
         status_code = error.response.status_code if error.response else "unknown"
-        logger.warning(
-            f"Couldn't load weather forecast from met.no (status: {status_code})"
-        )
+        logger.warning(f"Couldn't load weather forecast from met.no (status: {status_code})")
 
     except Exception as error:
         logger.error(error, exc_info=True)
